ml_service/transcription/requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
websockets>=10.0

# Audio processing
whisper-timestamped>=1.12.20
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.1
librosa>=0.10.0
pyannote.audio>=3.1.1

# Text processing
punctuator>=1.2
spacy>=3.7.0

# Utilities
pydantic>=2.0.0
python-dotenv>=0.19.0
tqdm>=4.65.0
requests>=2.31.0
